+++ CHANGES.rst
@@ -118,6 +118,9 @@ New Features
   - Added Earth radius (``R_earth``) and Jupiter radius (``R_jup``) to units.
     [#4818]
 
+  - Added a ``represents`` property to allow access to the definition of a
+    named unit (e.g., ``u.kpc.represents`` yields ``1000 pc``). [#4806]
+
 - ``astropy.utils``
 
   - ``Path`` object could be passed to ``get_readable_fileobj``. [#4606]